requirements.txt
pyyaml
pydantic>=2
transformers
datasets
accelerate
peft
trl
torch
tqdm
rich
jsonlines
scikit-learn
